fixup! refactor(upgrade): use Bazel packages to avoid symlinks in the source
