+++ packages/upgrade/BUILD.bazel
@@ -7,13 +7,13 @@ ng_module(
     srcs = glob(
         [
             "*.ts",
-            "dynamic/src/*.ts",
+            "src/dynamic/src/*.ts",
         ],
     ),
     deps = [
         "//packages/core",
         "//packages/platform-browser-dynamic",
-        "//packages/upgrade/common",
+        "//packages/upgrade/src/common",
         "@npm//zone.js",
     ],
 )
