+++ packages/upgrade/public_api.ts
@@ -12,7 +12,7 @@
  * Entry point for all public APIs of this package. allowing
  * Angular 1 and Angular 2+ to run side by side in the same application.
  */
-export {VERSION} from './common/src/version';
-export {UpgradeAdapter, UpgradeAdapterRef} from './dynamic/src/upgrade_adapter';
+export {VERSION} from './src/common/src/version';
+export {UpgradeAdapter, UpgradeAdapterRef} from './src/dynamic/src/upgrade_adapter';
 
 // This file only re-exports content of the `src` folder. Keep it that way.
