+++ packages/upgrade/src/common/src/component_info.ts
