+++ packages/upgrade/src/common/src/downgrade_component.ts
