+++ packages/upgrade/src/common/src/downgrade_component_adapter.ts
