+++ packages/upgrade/src/common/test/BUILD.bazel
@@ -3,17 +3,12 @@ load("//tools:defaults.bzl", "ts_library", "ts_web_test_suite")
 ts_library(
     name = "test_lib",
     testonly = True,
-    srcs = glob([
-        "**/*.ts",
-    ]),
+    srcs = glob(["**/*.ts"]),
     deps = [
         "//packages/core",
         "//packages/core/testing",
-        "//packages/platform-browser",
-        "//packages/platform-browser-dynamic",
-        "//packages/upgrade",
-        "//packages/upgrade/common",
-        "//packages/upgrade/common/test/helpers",
+        "//packages/upgrade/src/common",
+        "//packages/upgrade/src/common/test/helpers",
     ],
 )
