+++ packages/upgrade/src/common/test/component_info_spec.ts
