+++ packages/upgrade/src/common/test/downgrade_component_adapter_spec.ts
