+++ packages/upgrade/src/common/test/helpers/BUILD.bazel
@@ -9,6 +9,6 @@ ts_library(
     ]),
     deps = [
         "//packages/core/testing",
-        "//packages/upgrade/common",
+        "//packages/upgrade/src/common",
     ],
 )
