+++ packages/upgrade/static/BUILD.bazel
@@ -15,6 +15,6 @@ ng_module(
     deps = [
         "//packages/core",
         "//packages/platform-browser",
-        "//packages/upgrade/common",
+        "//packages/upgrade/src/common",
     ],
 )
