+++ packages/upgrade/static/src/downgrade_module.ts
@@ -9,9 +9,9 @@
 import {Injector, NgModuleFactory, NgModuleRef, StaticProvider} from '@angular/core';
 import {platformBrowser} from '@angular/platform-browser';
 
-import {IInjectorService, IProvideService, module as angularModule} from '../../common/src/angular1';
-import {$INJECTOR, $PROVIDE, DOWNGRADED_MODULE_COUNT_KEY, INJECTOR_KEY, LAZY_MODULE_REF, UPGRADE_APP_TYPE_KEY, UPGRADE_MODULE_NAME} from '../../common/src/constants';
-import {LazyModuleRef, UpgradeAppType, getDowngradedModuleCount, isFunction} from '../../common/src/util';
+import {IInjectorService, IProvideService, module as angularModule} from '../../src/common/src/angular1';
+import {$INJECTOR, $PROVIDE, DOWNGRADED_MODULE_COUNT_KEY, INJECTOR_KEY, LAZY_MODULE_REF, UPGRADE_APP_TYPE_KEY, UPGRADE_MODULE_NAME} from '../../src/common/src/constants';
+import {LazyModuleRef, UpgradeAppType, getDowngradedModuleCount, isFunction} from '../../src/common/src/util';
 
 import {angular1Providers, setTempInjectorRef} from './angular1_providers';
 import {NgAdapterInjector} from './util';
