+++ packages/upgrade/static/test/BUILD.bazel
@@ -12,8 +12,8 @@ ts_library(
         "//packages/platform-browser",
         "//packages/platform-browser-dynamic",
         "//packages/platform-browser/testing",
-        "//packages/upgrade/common",
-        "//packages/upgrade/common/test/helpers",
+        "//packages/upgrade/src/common",
+        "//packages/upgrade/src/common/test/helpers",
         "//packages/upgrade/static",
     ],
 )
