+++ packages/upgrade/static/test/integration/change_detection_spec.ts
@@ -11,9 +11,10 @@ import {async} from '@angular/core/testing';
 import {BrowserModule} from '@angular/platform-browser';
 import {platformBrowserDynamic} from '@angular/platform-browser-dynamic';
 import {UpgradeComponent, UpgradeModule, downgradeComponent} from '@angular/upgrade/static';
-import * as angular from '../../../common/src/angular1';
 
-import {html, withEachNg1Version} from '../../../common/test/helpers/common_test_helpers';
+import * as angular from '../../../src/common/src/angular1';
+import {html, withEachNg1Version} from '../../../src/common/test/helpers/common_test_helpers';
+
 import {bootstrap} from './static_test_helpers';
 
 withEachNg1Version(() => {
