+++ packages/upgrade/static/test/integration/downgrade_component_spec.ts
@@ -11,9 +11,9 @@ import {async, fakeAsync, tick} from '@angular/core/testing';
 import {BrowserModule} from '@angular/platform-browser';
 import {platformBrowserDynamic} from '@angular/platform-browser-dynamic';
 import {UpgradeComponent, UpgradeModule, downgradeComponent} from '@angular/upgrade/static';
-import * as angular from '../../../common/src/angular1';
+import * as angular from '../../../src/common/src/angular1';
 
-import {html, multiTrim, withEachNg1Version} from '../../../common/test/helpers/common_test_helpers';
+import {html, multiTrim, withEachNg1Version} from '../../../src/common/test/helpers/common_test_helpers';
 import {$apply, bootstrap} from './static_test_helpers';
 
 withEachNg1Version(() => {
